Resolved signage issues. Fixed casting problems with atoi by changing to (char + '0'). Reduced print frequency. Penalised network for using too many neurons with no progress. Changed output type to single for pi generation. This means only the neuron with the most powerful fire will be accepted as input at a time. Added some omp parallelisation. Unsure if actually working though.

+++ morassnetwork.cpp
@@ -5,14 +5,14 @@ MorassNetwork::MorassNetwork(void) {
 }
 
 MorassNetwork::~MorassNetwork(void) {
-	for (int m = 0; m < nodes.size(); m++) {
+	for (unsigned int m = 0; m < nodes.size(); m++) {
 		delete nodes[m];
 	}
 	nodes.clear();
 }
 
 void MorassNetwork::reset_charge() {
-	for (int i = 0; i < nodes.size(); i++) {
+	for (unsigned int i = 0; i < nodes.size(); i++) {
 		nodes[i]->reset_charge();
 	}
 }
@@ -24,11 +24,11 @@ std::vector<std::tuple<double, int>> MorassNetwork::step(void) {
 	return pulses;
 }
 
-void MorassNetwork::inject_node(int node_id, double input) {
+void MorassNetwork::inject_node(unsigned int node_id, double input) {
 	nodes[node_id]->inject_charge(input);
 }
 
-int MorassNetwork::length() {
+unsigned int MorassNetwork::length() {
 	return nodes.size();
 }
 
@@ -38,61 +38,61 @@ int MorassNetwork::add_node(double T, double Cp, double Cf, double Dp, double Df
 }
 
 bool MorassNetwork::del_node(int node_id) {
-	if (node_id < nodes.size()) {
-		for (int i = 0; i < nodes.size(); i++)
+	if (node_id < (signed)nodes.size()) {
+		for (unsigned int i = 0; i < nodes.size(); i++)
 			nodes[i]->del_edge(nodes[node_id]);
 		delete nodes[node_id];
 		nodes.erase(nodes.begin() + node_id);
 	}
 	else return false;
-	for (int i = 0; i < nodes.size(); i++)
+	for (unsigned int i = 0; i < nodes.size(); i++)
 		nodes[i]->set_id(i);
 	return true;
 	//if (nodes.size() > 0)
 	//	nodes[0]->set_counter(nodes.size());
 }
 
-bool MorassNetwork::add_edge(double out_pcnt, double out_fixed, int start_node, int end_node) {
+bool MorassNetwork::add_edge(double out_pcnt, double out_fixed, unsigned int start_node, unsigned int end_node) {
 	return nodes[start_node]->add_edge(out_pcnt, out_fixed, nodes[end_node]);
 }
 
-bool MorassNetwork::del_edge(int start_node, int end_node) {
+bool MorassNetwork::del_edge(unsigned int start_node, unsigned int end_node) {
 	return nodes[start_node]->del_edge(nodes[end_node]);
 }
 
 std::vector<int> MorassNetwork::remove_stranded_nodes() {
 	std::vector<Node*> no_outputs;
 	std::vector<int> to_remove;
-	int curr_node_id;
+	unsigned int curr_node_id;
 	bool stranded;
-	for (int i = 0; i < nodes.size(); i++) {
+	for (unsigned int i = 0; i < nodes.size(); i++) {
 		if (nodes[i]->get_num_edges() == 0) {
 			no_outputs.push_back(nodes[i]);
 		}
 	}
-	for (int i = 0; i < no_outputs.size(); i++) {
+	for (unsigned int i = 0; i < no_outputs.size(); i++) {
 		stranded = true;
 		curr_node_id = no_outputs[i]->get_id();
-		for (int j = 0; j < nodes.size() && stranded; j++)
+		for (unsigned int j = 0; j < nodes.size() && stranded; j++)
 			if (curr_node_id != j) stranded = !nodes[j]->has_edge(no_outputs[i]);
 		if (stranded) to_remove.push_back(curr_node_id);
 	}
-	for (int i = to_remove.size()-1; i >= 0; i--) { //Delete in reverse order as the ids get updated after deletion
-		del_node(to_remove[i]);
+	for (int i = (int)to_remove.size()-1; i >= 0; i--) { //Delete in reverse order as the ids get updated after deletion
+		del_node(to_remove[(unsigned)i]);
 	}
 	return to_remove;
 }
 
 void MorassNetwork::print_network(void) {
-	for (int i = 0; i < nodes.size(); i++) {
+	for (unsigned int i = 0; i < nodes.size(); i++) {
 		nodes[i]->print_edges();
 	}
 }
 
 std::vector<std::tuple<double, int>> MorassNetwork::process_thresholds(void) {
 	std::vector<std::tuple<double, int>> pulses;
-	for (int i = 0; i < nodes.size(); i++) pulses.emplace_back(0.0, i);
-	for (int i = 0; i < nodes.size(); i++) {
+	for (unsigned int i = 0; i < nodes.size(); i++) pulses.emplace_back(0.0, i);
+	for (unsigned int i = 0; i < nodes.size(); i++) {
 		if (nodes[i]->over_threshold())
 			std::get<0>(pulses[i]) = nodes[i]->pulse(nodes[i]->trigger());
 	}
@@ -104,13 +104,13 @@ Node* MorassNetwork::get_node(int node_id) {
 }
 
 void MorassNetwork::process_decay(void) {
-	for (int i = 0; i < nodes.size(); i++) {
+	for (unsigned int i = 0; i < nodes.size(); i++) {
 		nodes[i]->decay();
 	}
 }
 
 void MorassNetwork::process_inputs(void) {
-	for (int i = 0; i < nodes.size(); i++) {
+	for (unsigned int i = 0; i < nodes.size(); i++) {
 		nodes[i]->assimilate_charge();
 	}
 }
